Improve Windows installation script
Since the default execution policy for PowerShell is Restricted, the
only way to run a PowerShell script in most environments is to pipe the
script into the PowerShell command.  Signing the script is insufficient.
It's possible a sophisticated anti-virus program will detect this and
prevent it.  In that case, the work-around is for developers to install
the prerequisites manually before running the configuration script.